Split MCP OAuth scope tests
